fix: resolve 4 clippy lints in webhook API
- io_other_error: use std::io::Error::other() shorthand
- manual_range_contains: use (400..500).contains()
- unnecessary_cast: remove u32-to-u32 casts (turn_count, edit_count)

+++ crates/server/src/webhook_engine/formatters/mod.rs
@@ -103,12 +103,12 @@ pub fn build_payload(
                 }
             },
             turn_count: if session.hook.turn_count > 0 {
-                Some(session.hook.turn_count as u32)
+                Some(session.hook.turn_count)
             } else {
                 None
             },
             edit_count: if session.jsonl.edit_count > 0 {
-                Some(session.jsonl.edit_count as u32)
+                Some(session.jsonl.edit_count)
             } else {
                 None
             },
